tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-play-state_default.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-play-state default</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-play-state-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="animation-play-state does not inherit by
   default">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         from {
            left: 0px;
         }
         to {
            left: 50px;
         }
      }
      div {
         width: 50px;
         height: 100px;
         background: red;
      }
      div > div {
         background: blue;
         position: relative;
         animation-duration: 1s;
         animation-timing-function: step-end;
         animation-fill-mode: forwards;
      }
      .test, .test > div {
         animation-name: move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var divs = document.querySelectorAll('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var outerStyle = getComputedStyle(divs[1], null);
         var innerStyle = getComputedStyle(divs[2], null);
         var passed = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         setTimeout(function() {
            divs[1].style.setProperty('animation-play-state', 'paused', '');
            passed = (outerStyle.left == '0px' && innerStyle.left == '0px');
         }, 500);
         divs[2].addEventListener('animationend', function() {
            clearTimeout(timeout);
            var state = innerStyle.getPropertyValue('animation-play-state');
            passed = (passed && state == 'running' && outerStyle.left == '0px'
                      && innerStyle.left == '50px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         divs[1].className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue block below should abruptly change from a tall rectangular shape
   to a square shape, and there should be no red.</p>
   <p>Testing...</p>
   <div>
      <div>
         <div></div>
      </div>
   </div>
</body>
</html>
